requirements.txt
langchain==0.0.225
ctransformers==0.2.5
sentence-transformers==2.2.2
pypdf
flask
faiss-cpu
python-dotenv
-e .